Rename state file
