+++ src/components/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Display from './Display';
 import ButtonPanel from './ButtonPanel';
 import calculate from '../logic/calculate';
-import { resetState } from '../state/utils';
+import { resetState } from '../state/state';
 
 class App extends React.Component {
   constructor(props) {
